refactor: move `Logging` to `logging` subpackage
