+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/core/CoreRegistrant.kt
@@ -6,6 +6,7 @@ import com.qualcomm.robotcore.eventloop.opmode.OpMode
 import com.qualcomm.robotcore.eventloop.opmode.OpModeManagerImpl
 import com.qualcomm.robotcore.eventloop.opmode.OpModeManagerNotifier
 import org.firstinspires.ftc.ftccommon.external.OnCreateEventLoop
+import org.firstinspires.ftc.teamcode.core.logging.Logging
 
 /**
  * This object is responsible for registering and running the core systems required for all opmodes.
